requirements/common.txt
# Core Stuff
# -------------------------------------
Django==2.1.2

# Configuration
# -------------------------------------
argon2-cffi==18.3.0
django-environ==0.4.5
django-sites==0.10
python-dotenv==0.9.1
django-cors-headers==2.4.0



# Extensions
# -------------------------------------
pytz==2018.5

# Models
# -------------------------------------
psycopg2-binary==2.7.5

Pillow==5.3.0
django-extensions==2.1.3
django-uuid-upload-path==1.0.0
django-versatileimagefield==1.10

# REST APIs
# -------------------------------------
djangorestframework==3.8.2
django-rest-swagger==2.2.0
drf-nested-routers
django-extra-fields


# LOGGING
# -------------------------------------
django-log-request-id==1.3.2

# Documentation
# -------------------------------------
mkdocs==1.0.4

# Raven is the Sentry client
# --------------------------
raven==6.9.0

# Async Tasks
# -------------------------------------
celery[redis]==4.2.1

# Auth Stuff
# -------------------------------------
PyJWT==1.6.4
django-mail-templated==2.6.5
